Refine parameters and thresholds
Utilize min_samples in DBSCAN to ignore clusters that are unlikely
to be cells given their small size. With this change, we can rely
more on Otsu's method for the initial image and remove the 2.5x
factor for the threshold. Accordingly, we can refine and reduce
the sizing of the bounding boxes since the detected cells will be
more accurately sized.

Additionally, introduce coordinate bounds for computing the boxes
to ensure that the boxes do not extend outside of the image.

This also fixes a bug where noisy samples (labeled as -1) were
included as a cluster.