Remove border elements prior to finding new centroids
Define remove_border_elements which removes components that are
touching the border (i.e. bright at the edge of the image). Do this
only if the component is less than a certain fraction of the total
components - this protects the primary cell if the bounding box is
too small.

This update does not change the calculated centroids for most cells
but shows marked improvement for a select few. This should also
considerably improve the updated bounding boxes, though this is not
reflected in the current notebook as the displayed boxes are from
the initial calculations. With this change, we also remove the
hacky factor used to enlarge the initial bounding box, with no loss
of accuracy in the updated centroids.