Fix bounding box size and apply Canny to whole image
We previously were updating the bounding box size with each
iteration, resulting uncontrolled expansion of the box as new cells
enter the frame. Establishing a fixed box size with the initial
image addresses this issue. This assumes that the size of the cell
will not substantially change over iterations.

We were also previously applying Canny to each sub-image -
convolving and applying gradients to tiny images is suboptimal,
especially when there's no reason not to work with the whole image
instead.

This markedly improves tracking performance and we observe that
large cells are more accurately tracked. Next steps would be to
address cells that 'merge' - currently, the centroids clump
together and moves in a group on top of a single cluster.

With this update, we also clean up our notebook a bit to remove
redundant and over-engineered code.