Overlap nearby centroids
As previously mentioned in 85faacf, some cells will clump together
and travel in groups. Our previous bounding box tracking will lose
effectiveness in this situation, as multiple cells appearing in one
bounding box will result in inaccurate centroids.

With this update, we conservatively identify cells that are
clustered together - namely, cells with centroids that fall in each
others' bounding boxes. Note that for cells a, b, and c to be
clustered together, a must overlap with b and c, and b must overlap
with a and c, and c must overlap with a and b (i.e. a complete
overlap). For these cells, update their centroids to be the mean,
and update their bounding boxes to be the max (i.e. combine the
bounding boxes and find the max edge distance from the new
centroid).

One notable issue is that there are occasional cells that will bump
into a group but not attach, and are subsequently lost to tracking
when their centroids are merged with the group despite their
eventual separation. This should be addressed in a later update.