Fill cells after dilation and erosion
Previously, we identified centroids using cell edges by applying
the Canny edge detector to each image. This weakens the signal for
large cells as we no longer account for the volume that each cell
occupies. We address this now by first dilating and eroding the
image to close gaps in the edges (i.e. we close the image), then
filling the image and thereby the cells. Subjectively, cell
tracking appears to be substantially improved, with more accurate
centroids and more conservative cell merging.

Our code is now a little more complicated but there is now a clear
image processing pipeline. Thus, let's rearrange the notebook so
that core functions are organized more intuitively.