Remove secondary components from bounding boxes
We previously removed border components from each bounding box to
improve tracking accuracy, as detailed in 9349bb1. We can extend
this to remove all secondary components - that is, keep only the
largest component which we assume to be the cell, and remove all
others. This change has predictable effects - previous bounding
boxes that had smaller bits and pieces in addition to the primary
cell will now have more accurate centroids. An added benefit to
this is that bounding boxes will now expand less rapidly, as the
boxes will no longer grow aggressively whenever something floats
into the frame.