Redraw bounding boxes
With the improvements in 4769731 and 2b02174, we are now able to
accurately identify the primary cell within bounding boxes.
Isolating the cell allows us to actively redraw the boxes, relative
to the size of the cell. This improves the accuracy of each box,
and thus we can reduce the sizing factor to 1.2 (from 1.5).