Only merge cells if they are connected
Previously we would merge cells if their centroids were within each
others' bounding boxes, as detailed in 53a364d. This led to
occasional premature merges, particularily in cases with large
bounding boxes. With this update, we only merge two cells/boxes if
a singular connected component can be identified in the resulting
(larger) box. We can keep the prior logic for overlapping bounding
boxes as a trigger to check for a merge event, and we can use
existing code to isolate a primary cell. Note that we use a more
conservative threshold for component removal here.