Color distance graph using merge events
In line with the way we color bounding boxes, color the lines in
our cumulative distance plot using the final number of merge events
for the cell. Similarly, we weigh the merged lines a little heavier
than unmerged ones.

It would be nice if we could change colors within a line as merge
events occur, but let's table that for another day.