Extract features with pyradiomics and cluster using kmeans
As a rough, initial implementation of cell classification, we
extract features from grayscale images using pyradiomics defaults
and perform kmeans clustering on those features. Presently at this
point we do not perform any form of dimensionality reduction.

Creation of masks for cells is done using Otsu's thresholding to
binarize a Gaussian blurred image, followed by dilation, erosion,
and filling to create connected components which we isolate as
individual cells. We ignore masks that are below a certain size
threshold.

Of note, in order to use pyradiomics we downgrade our Python
version from 3.9 to 3.7. This does not break or change our
celltrack notebook. Update requirements.txt accordingly for binder
support.