Classify adherent and suspension images
Now that we have a baseline classifier working, fit it on the
control image and have it classify cells in the adherent and
suspension images.

Enable shape2D and disable SumAverage in the feature extractor, as
previously recommended by warnings.

Of note, we currently do not scale the features - it seems to
greatly mess with the classification results. Will need to
reevaluate this at a later time.