Determine n_clusters using Silhouette score
We test varying n_clusters and plot both inertia and silhouette
scores to determine the number of clusters to use for kmeans. We
see that using the elbow method we could justify 7 or 9 clusters,
though from testing this produces a few exceedingly small (<5)
clusters. Let's use the global max of the silhouette score, which
leads to 2 clusters (and thus the same output as before).