Preprocess images with histogram equalization
Our previous approach of clipping and redistributing pixel values
in 4354e28 seemed to negatively impact detection in the adherent
images, so let's use histogram equalization for preprocessing
instead. This does create some noise in the image, so apply a
denoiser to reverse that effect. Adjust the number of clusters to
4 as guided by manual/visual inspection.