Enable contrast enhancement with adaptive thresholding
As an alternative to histogram equalization, let's attempt to use
contrast enhancement to separate cells from the background. We do
this by pushing pixel values away from the peak histogram value of
the background.

Combined with adaptive thresholding (this performs poorly with
Otsu's), this seems to perform similarly to our previous approach
of histogram equalization followed by Otsu's, with tradeoffs in
masking performance and notable differences in final labels. Let's
push these changes so we can compare.