Invert image and subtract background
In this commit we invert the image and aggressively subtract the
background, clipping negative values. By doing this we sacrifice
details in the bright regions so that we can then aggressively
increase contrast in the dark regions - this allows us to extract
a little more information from the very faint and flat adherent
cells. This is a tad bit noisier of a process though, thereby
jiggling centroids around a bit, but our masks do demonstrably
improve in several regions.

Let's also experiment with applying multi-Otsu on a clipped
histogram - updating our version of skimage allows us to do this
easily.