Quantify and visualize fluorescence within clusters
In order to assess variation in fluorescence across clusters, let's
count the number/proportion of fluorescent pixels in each cluster
and visualize the area that is fluorescent.

This takes a binary approach to fluorescence - a pixel is either
fluorescent or not. We can also consider quantifying intensity in
a later update.

Let's also update our PyYAML version to 6.0, to address a security
vulnerability in 5.3.1.