Use new thresholded and background-subtracted images
Add adherent_color-1.tiff, which is a pre-processed version of
adherent_color.tiff. This image stack on visual inspection does
appear to have improved definition of cells and more prominent
fluorescence so let's use it for our analysis, though in theory
it is just a substitution of our own pre-processing steps. We'll
update the offset in params_gray_mask.yml to account for the
new background, using the same process as before which is to
lower the value until noise is visually acceptable. We also re-
enable Canny unmasking on the grey image, since the improved
background separation does appear to highlight more out-of-focus
regions.

A few other changes in this update:
- Add display_hist to displaytools.py
- Update Canny unmasking to support custom thresholds
- Use identical params for image_fluorescent and
  flourescence_binarized (but skip the post-processing steps for
  the latter since we're attempting to quantify true fluorescence)