Run k-means on full image stack
Let's cluster/classify the full time course captured by the image
stack using the same process established prior. We'll only use the
first image for fitting the scalers/classifiers and fix k=4 for
now, though we should consider determining the number of clusters
and/or classifying each image independently, as new information may
be introduced with time.

We can then plot the counts of each class and observe how the
distribution changes over time.

In this update we change the offset for fluorescent mask extraction
from -3 to -12, to account for masks growing excessively large in
the later images. A more optimal solution may be to shift the
offset for each image based on comparison to the initial image, but
we'll save that for later work.