Use suspension image stack
We have a new image stack that contains larger and a greater
quantity of images, so use that for our analysis. There are now
two sets of fluorescent images for each time point, so let's
disable our fluorescent analysis for now and revise later. We
actually don't need the fluorescent images for masking in this
case because the cell clusters seem to be quite well separated,
and the fluorescence doesn't seem to highlight the flat thin cells
all that well either.

We'll update our masking process to use adaptive thresholding -
note that this image stack does not work well with manual
thresholding since there seems to be a significant gradient in
terms of darker and lighter areas. We also add new functionality
to trim borders of each image, since it appears that some of the
images have increased fluorescence/details at the edges that
appear to be just noise.