Visualize fluorescence trends over time
We can visualize how fluorescence changes for each class using
heatmaps. Note that the max intensity of the fluorescence only
goes up to ~75, so we don't scale the image (or the heatmap will
become patchy). The frequency is much higher at low intensities
(namely 0) so let's use a log norm to straighten out the
distribution.

We can also characterize the change using an arbitrary percentile -
we plot the 99th percentile of pixel intensity at each given time
point for each given class.