Revise size unmasking and update params
We previously unmasked based on region size prior to dilation and
erosion, but it probably makes sense to switch the order - small
areas next to big areas are more likely to be hints of very flat
and subtle adherent cells when compared to small regions in
isolation. With this change we can remove the size unmasking logic
in extract_features since this is automatically handled by masking.

We'll also update our params now to be a little more precisely
calibrated - use a closure kernel size of 8 so that we don't
aggressively overlap regions, and unmask up to a size of 250 pixels
since at this point it seems like all meaningful areas are included
except for that one barely noticeable adherent cell.