Plot total and mean areas of classifications
The heatmaps show raw counts of pixels for each class which can be
a bit misleading if the classes don't have similar numbers of
pixels. Let's plot the total area (in pixels) over time so we have
a point of reference. We can also plot the mean cell (mask region)
area as well.

As an extension of e7a7033, we no longer need to pass around
`mask_image` since it is already encapsulated by `mask_labels`.
We can thus clean up our method signatures accordingly.