Estimate feature importance using LogisticRegression
Though k-means in itself does not offer information on feature
importance, we can use the labels it generates for supervised
learning. This involves transforming the labels for each class (1
if within the class and 0 if not) and fitting a logistic regression
model using the scaled (pre-PCA) dataframe, giving us coefficients
for each feature for each class.