Parameter search proof of concept
Given inherent background variation across each image stack we
should implement a way to intelligently adjust model parameters as
we step through the images. In particular, the offset parameter is
a good target for tweaking as the background can be a bit brighter
or darker depending on the image.

In this proof of concept, we sweep through a search space and track
mask image output with two metrics: the total area initially
captured and the number of unique cell regions that is in our final
mask output. We would ideally like to mask as much as possible
before the total area sharply increases (i.e. we start picking up
background) or the number of cell regions sharply decreases (i.e.
there isn't enough separation between unique regions).

We observe a clear uptick in total area as we increase the offset
though the number of cell regions appears to be more stable. Total
area thus may be a more useful tool for detecting appropriate
parameters.