Modify Canny unmasking logic and fix get_mask_image bug
Our prior method for unmasking via Canny edges was built on an
assumption that intensity-detected edges would be a strict subset
of a filled Canny image, which is a bit of a crude approach - there
are likely regions where edges are sparsely detected and thus by
&-ing the two images together we are more likely to get a patchier
mask.

Let's rework our approach to unmask after image closure, in the
same step as the size thresholding, and remove components if and
only if no edges are detected within the component.

Fix a get_mask_image variable naming bug, introduced in 99940d8.

This commit has been tested but is not reflected in the notebook.