Refine masking offset for each image
In continuation of 99940d8, let's sweep through possible offsets
for all images in the image stack. This should give us more
generous coverage of the regions of interest. We also continue to
perform Canny unmasking following the update in 3e71f67, but we can
look into tweaking the Canny parameters as well as the closure
kernel size in a later update.