Limit rows when fitting kmeans
A consequence of 03b9d7f is that we can have a prohibitively large
number of data points to fit with (e.g. 500k+ for four stacks of
100 images each). Let's downsample to 40k rows prior to fitting,
which gives us a reasonable runtime for finding silhouette scores.

We could theoretically use more rows, especially when fitting the
final model, but 40k should be plenty for identifing fewer than
10 clusters...