Standardize output pathing
In this change we stop using __temp files for generated output.
Instead, scripts can take a path to an output directory and create
timed entries for each image stack of interest. Generally speaking,
the directory structure is:
/[output directory]/[image stack]/[timestamp]/[output file(s)]

For output specific to individual image stacks, timestamps are
propagated from mask image creation - subsequent scripts will look
for the latest generated mask images.

Since kmeans output does not run on a similar timeline (we can fit
models using several image stacks), the pathing is a bit different:
/[output directory]/[kmeans/silhouette_scores]/[timestamped file]

Note that this code is presently untested as it is intended to be
run on Quest.