Revise offset selection using mask diffs
Our previous approach implemented in c7611fb made an inappropriate
assumption - that acquired background would first appear in the
largest region. This is not true, as we see many examples of
background appearing in regions dwarfed by the largest components.

Let's try a new approach - as we increase our offset, we identify
new pixels included in the mask and check the size of the largest
connected clump of new pixels. Increased coverage around a cell
region of interest is unlikely to appear as a large new region, so
hopefully this will identify swaths of background that get picked
up. We set an arbitrary cutoff of 2000 based on preliminary
findings.