Recenter background when scaling images
We previously scaled images by stretching the maximum intensity to
255. First, let's also make sure that the minimum intensity is 0.

This approach notably does not standardize the background intensity
which is problematic when image stacks contain fluctuating
background intensities. We now re-center the mode of every image to
a predefined value, using the mode of the first image, which is
roughly equivalent to re-centering the background.