Center images on arbitrary common mode
We previously centered images on the mode of the first image of
each stack. This is problematic because some image stacks are much
darker and therefore become centered in a way that is very
different from other image stacks, resulting in notably different
classification output. Let's instead center on an arbitrarily
chosen intensity, 90, based on the modes of our current image
stacks of interest.