Check true mask diff when refining offset
As an addendum to 34af871, let's also check the true mask diff when
determining an appropriate offset. We previously checked a more
generous closure since at times the true mask would under-cover
regions, but if we don't see too much coverage with the true mask
we can probably keep advancing as well. So, check both the true and
test masks for stopping criteria. Let's increase our arbitrary
cutoff from 2000 to 3000 as well, to be a bit more generous with
coverage.