Speed up and validate offset refinement process
This update addresses two major issues we noticed during offset
refinement: extremely lengthy runtimes and errors where a max size
component could not be found.

The issue with lengthy runtimes stemmed primarily from the
unmasking step - in images with tens of thousands of components we
spent a lot of time iterating through each one. We now filter on
size using much more elegant numpy calls, and then separate the
Canny unmasking step in order to re-identify (fewer) components,
all in all resulting in significantly fewer array operations and
much improved runtime. On testing it seems that we can also
increase our size threshold a touch, to 60 pixels, which will
further reduce the number of regions to step through.

On a related note, it seems that with denoising set to 3 we still
produce a noisy image with many tiny components - on manual
testing increasing this value to 5 seems to produce better results.

In order to address the np.max error, we need to check for three
bad scenarios: 1) nothing is masked, 2) everything is masked, 3)
the new mask is identical to the previous mask. For 1) and 2), we
add a validation step, where, if we fail validation and have
previously identified a working mask, we stop and return the
working mask. In this validation step, we can also perform a
sanity check: if the largest component is greater than 10% of the
image, we probably masked too much. We'll set a cutoff of 15%,
though, to be on the conservative side. To address 3), we now check
to make sure that there is indeed change in the mask diffs before
attempting to find the largest components.

Lastly, reorder parameter ordering in the YAML files to match the
order of operations in get_mask_image.