Lower diff cutoff, exclude redundant CSVs, clean up print output
Revert the diff size cutoff to 2000, which was increased in 5cc49e8
and noted to cause issues when tested with 1A1. Basically, it is
possible for the test mask to grow disproportionately faster than
the true mask (so the test mask cutoff can probably be higher), and
validation will fail on the test mask before the true mask criteria
was reached. Subjectively speaking, we ideally do not want to pick
masks based on the validation criteria, as it's more of a backup
option.

Also make a few minor revisions - make print output for contouring
more intuitive, and don't include redundant CSVs in the packaged
output.