Make improcessing.apply_imfill more robust
Previously we floodfilled our image from the four corners, which
encountered problems when the four corners were enclosed by regions
of interest. Add four more floodfill points (at the midpoints of
all edges) to avoid this issue. Another way to approach this is to
randomly select points to floodfill from, but this runs the risk of
not filling a region of interest. We could also consider removing
the border after closure, to ensure an open path towards the center
but this means that corner regions would not reliably be filled.