Increase offset search range and improve mask selection criteria
Lower the start of our offset search by 5 (-20->-25), since a very
small number of images (~1%) would benefit from it.

Validate test masks with a different cutoff (15%->40%) since the
aggressive closure is very likely to produce large components.

Set a hard 10,000 pixel cutoff for the true mask diff, since if the
true mask is growing by such a large amount the test mask is
irrelevant. Increase the cutoff for the test mask diff (2000->4000)
since the lower threshold was triggered a bit too easily.