Fix inconsistent mask binarization
This fixes a major bug with array handling in get_mask_image. A few
of our array operations produced non-binarized (0-255) images, such
as with the thresholding step, which is at best innocuous but at
worst alarmingly incorrect. Of note, update apply_imfill to fill
with 1 instead of 255, and use a more deliberate != 0 to flip bits
instead of cv2.bitwise_not, since the latter seems to produce 254
and 255 values for whatever reason. With this update, we should be
working exclusively with binary (0 or 1) mask images every step of
the way.

Also update the test diff size criteria to 5000 from 4000, and bump
test diff closure size to 5 from 4, since this seems to improve
results.