Purge redundant data.csv at end of run
We store a data.csv as a sanity check to corroborate with
data_labeled.csv - otherwise it is completely redundant. We really
don't need to create this CSV at all since the data should also be
stored in the pickle files, but since it's relatively innocuous to
generate let's just leave it be in case anything fails midway and
we need something easily human-readable. It's not a small file
though, so let's purge it at the end of each run.