Save fluorescence data to CSVs
Let's save a CSV file with our feature data, labels, and
fluorescence data (namely the area encompassed by the fluorescent
pixels, using Otsu's as a threshold). We are interested in the
total area of each mask region as well, and it is already encoded
as original_shape2D_PixelSurface within the feature data.

Let's also organize our output directory structure a little better.