Add percentile cutoff for fluorescence binarization
We previously binarized our fluorescent images using Otsu's, which
picks a cutoff of around 20 (on a 255 scale). On visual inspection,
lower cutoffs produce a significant amount of noise starting from
the image corners.

Let's try binarizing with a higher cutoff, which we can do using
a percentile cutoff. Setting the percentile to 99.5 generates a
cutoff of 50-100, depending on the image stack.