Save count data for fluorescent regions
As an extension of e527cea, which ignored the fluorescent data and
counted the number of regions using the brightfield images alone,
let's overlay our fluorescent images and count the number of
fluorescent regions, where we consider a region to be fluorescent
if there is at least one fluorescent pixel within the region. We'll
use the higher cutoff of the 99.5th percentile since that should
better highlight differences between clusters.