Add script for output validation
We occasionally encounter node failures on Quest in which jobs get
stuck (there have been several instances of jobs running for a few
seconds and freezing in the "CG" state). This script checks whether
the expected output files are present, and if not will output the
path of the image stacks that need to be re-run.